+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change default orchestrator and replicator credentials to overwrite-me/overwrite-me
 - Changed process of determining the healthy publish instance to check component init state instead of just AEM health
 - Change SSL Verification of the AEM SSL certificate to false #51
+- Change default AEM connection configuration from HTTP to HTTPS #54
 
 ## [2.0.0] - unknown