+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Change default orchestrator and replicator credentials to overwrite-me/overwrite-me
 - Changed process of determining the healthy publish instance to check component init state instead of just AEM health
+- Replaced ELB Client with ELBv2 client to support the new AEM OpenCloud architecture which uses an Application Load Balancer #47
 - Changed `com.shinesolutions.aemorchestrator.config.ProxyConfig` to return empty string instead of `null` #49
 - Change SSL Verification of the AEM SSL certificate to false #51
 - Change default AEM connection configuration from HTTP to HTTPS #54